simplified shooter functions
